apps/re-agent/requirements.txt
pydantic
gradio
poethepoet
python-dotenv
requests
langchain-google-genai
langchain-core
langgraph
langchain-community
langchain-core